improved handleHiscores.ts
